refactor : 공통 테스트 모듈로 활용하기 위해 AcceptanceTest 클래스명 BaseTest로 변경
