+++ src/test/java/nextstep/subway/BaseTest.java
@@ -1,0 +1,28 @@
+package nextstep.subway;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.web.server.LocalServerPort;
+
+import io.restassured.RestAssured;
+import nextstep.subway.utils.DatabaseCleanup;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class BaseTest {
+	@LocalServerPort
+	int port;
+
+	@Autowired
+	private DatabaseCleanup databaseCleanup;
+
+	@BeforeEach
+	public void setUp() {
+		if (RestAssured.port == RestAssured.UNDEFINED_PORT) {
+			RestAssured.port = port;
+			databaseCleanup.afterPropertiesSet();
+		}
+
+		databaseCleanup.execute();
+	}
+}
